dev-requirements.txt
django~=4.0.0
django-environ==0.9.0
djangorestframework
whitenoise==6.2.0
gunicorn==20.1.0